requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
streamlit>=1.40.0
openai>=1.0.0
pinecone>=7.0.0
pypdf>=5.0.0
python-docx>=1.1.0
python-dotenv>=1.0.1
requests>=2.32.0

